+++ packages/jsii-dotnet-runtime/src/Amazon.JSII.Runtime/JsiiException.cs
@@ -1,11 +1,15 @@
-﻿using Amazon.JSII.JsonModel.Api.Response;
-using System;
+﻿using System;
+using Amazon.JSII.JsonModel.Api.Response;
 
 namespace Amazon.JSII.Runtime
 {
     public class JsiiException : Exception
     {
         public ErrorResponse ErrorResponse { get; }
+
+        public JsiiException(string message) : base(message)
+        {
+        }
 
         public JsiiException(string message, Exception innerException)
             : base(message, innerException)