+++ packages/jsii-java-runtime/pom.xml.t.js
@@ -80,7 +80,7 @@ process.stdout.write(`<?xml version="1.0" encoding="UTF-8"?>
         <dependency>
             <groupId>javax.annotation</groupId>
             <artifactId>javax.annotation-api</artifactId>
-            <version>[1.2,)</version>
+            <version>[1.3.2,)</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>