+++ packages/jsii-pacmak/test/expected.jsii-calc-base/dotnet/Amazon.JSII.Tests.CalculatorPackageId.BasePackageId/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using Amazon.JSII.Runtime.Deputy;
 
-[assembly: JsiiAssembly("@scope/jsii-calc-base", "0.7.10", "scope-jsii-calc-base-0.7.10.tgz")]
+[assembly: JsiiAssembly("@scope/jsii-calc-base", "0.7.13", "scope-jsii-calc-base-0.7.13.tgz")]